requirements.txt
matplotlib
mpl-toolkits.clifford
numpy
scipy
